requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
jsonschema>=4.19.0
pyperclip>=1.8.2